Add 'record size' parameter
There's now a "record size" parameter, that will automaticly
configure the expected values for diameter, innerRad and outerRad.

The previous values are still the default (for the default record
size, which is 12"), but there is already support for 5" records too.
Adding other sizes (7" and 10" are probably the most sought ones that
are still not supported) are just a matter of calculating what should
the values be for these sizes, and add an if clause with them.