create "groove and printer" profiles
In order to easilly develop the script in each project, or by each
person, and to use different printers and setups, this patch
introduces the concept of "profiles". In this way, one can easilly
add his own printer characteristics, or use one of the ones available
by just changing one value.